Cloud-Run-Python/requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
defusedxml==0.7.1
ecdsa==0.19.0
Flask==3.0.0
fonttools==4.55.3
fpdf2==2.7.6
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.3.1
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.62.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.17.0
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.0.1
